BUG Check for None
This happens if the message is moved while the Check Mail action is running

+++ rbit/sync.py
@@ -204,13 +204,14 @@ def update_folder(client, folder, create_session=None):
         changed = client.fetch_flags_since(prevmodseq)
         for uid in changed:
             m = load_message(account, folder, uid, lambda: session)
-            rfs = set(changed[uid]['FLAGS'])
-            lfs = set(f.flag for f in m.flags)
-            for f in m.flags:
-                if f.flag not in rfs:
-                    session.delete(f)
-            for f in rfs-lfs:
-                session.add(models.Flag(mid=m.mid, flag=f))
+            if m is not None:
+                rfs = set(changed[uid]['FLAGS'])
+                lfs = set(f.flag for f in m.flags)
+                for f in m.flags:
+                    if f.flag not in rfs:
+                        session.delete(f)
+                for f in rfs-lfs:
+                    session.add(models.Flag(mid=m.mid, flag=f))
         session.commit()
 
     # We need to wait until we have done all the updates to update the
